requirements.txt
flask 
flask-login 
mastodon.py 
pillow 
numpy 
open-clip-torch
